[hipfft] use templated common data-layout-related helpers in hipfft
